+++ src/routes/(app)/areas/[...slugs]/_/blocks/[blockSlug]/routes/[routeSlug]/edit/EditRoutePage.svelte
@@ -2,6 +2,7 @@
   import { enhance } from '$app/forms'
   import { page } from '$app/state'
   import { PUBLIC_APPLICATION_NAME } from '$env/static/public'
+  import { checkRegionPermission, REGION_PERMISSION_DELETE } from '$lib/auth'
   import AppBar from '$lib/components/AppBar'
   import DangerZone from '$lib/components/DangerZone'
   import FormActionBar from '$lib/components/FormActionBar'
@@ -58,4 +59,6 @@
   <FormActionBar label="Update route" pending={updateRoute.pending} />
 </form>
 
-<DangerZone name="route" onDelete={() => (route.id == null ? undefined : deleteRoute(route.id))} />
+{#if checkRegionPermission(pageState.userRegions, [REGION_PERMISSION_DELETE], route.regionFk)}
+  <DangerZone name="route" onDelete={() => (route.id == null ? undefined : deleteRoute(route.id))} />
+{/if}
